requirements.txt
apturl==0.5.2
asgiref==3.3.1
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.14.2
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==3.1
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
dicttoxml==1.7.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.4
djangorestframework==3.12.2
docopt==0.6.2
entrypoints==0.3
facturacion-electronica==0.5.4
future==0.18.2
httplib2==0.14.0
idna==2.8
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
lxml==4.5.2
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
netifaces==0.10.4
num2words==0.5.10
oauthlib==3.1.0
olefile==0.46
packaging==20.3
pdf417gen==0.7.1
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.24.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.1
suds-jurko==0.6
systemd-python==234
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.10
wadllib==1.3.3
xkit==0.0.0
XlsxWriter==1.3.3
xmltodict==0.12.0
zim==0.72.0
